theme: trying to move into a dark theme (WIP)
the config settings to set colors in the theme are not working well. first of all it should be foreground and background colors instead of black and white so we can switch between colorschemes better. secondly there is lots of hard coded coloring going on that prevents simply changing colors via config.